Fixed hotkey window issue.
As previously configured, whenever I had a fullscreen iTerm2 window and
I enabled the hotkey window, the first row of the hotkey window would be
obscured by the menu bar at the top of the screen. Now, I've set the
hotkey window to always show up in Space 1, which fixes the issue.